cDAC DBI: fix x86 stack overrun in ParseContinuation (async stackwalk)
ICorDebugProcess12::GetAsyncStack -> CordbAsyncStackWalk::PopulateFrame calls
IDacDbiInterface::ParseContinuation, whose native ABI declares the out-param as
PCODE* pDiagnosticIP. PCODE == TADDR is host-pointer-sized (4 bytes on x86, 8 on
x64), and the RS caller passes the address of a 4-byte PCODE stack local. The
cDAC modeled the parameter as ulong* and wrote 8 bytes through it, overrunning
the caller's stack slot by 4 bytes on x86 and tripping the /GS stack cookie
(STATUS_STACK_BUFFER_OVERRUN, 0xC0000409). On x64 PCODE is 8 bytes so the write
fit, which is why the async tests passed on x64 but crashed on x86.

Model pDiagnosticIP as nuint* to match the native PCODE* width, cast the written
value to nuint, and update the DEBUG cross-check local to nuint.

Validated on Windows x86 Debug (DEBUG cDAC/legacy cross-check active): all 7
previously-crashing Async tests now pass (AsyncBreakpoint, AsyncBreakpointJmc,
AsyncGeneric, AsyncRecursive, AsyncSharedGeneric, AsyncSimple,
AsyncV2CallingAsyncV1Iterator).

+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Legacy/Dbi/DacDbiImpl.cs
@@ -5403,7 +5403,7 @@ public sealed unsafe partial class DacDbiImpl : IDacDbiInterface
         return hr;
     }
 
-    public int ParseContinuation(ulong continuationAddress, ulong* pDiagnosticIP, ulong* pNextContinuation, uint* pState)
+    public int ParseContinuation(ulong continuationAddress, nuint* pDiagnosticIP, ulong* pNextContinuation, uint* pState)
     {
         int hr = HResults.S_OK;
         try
@@ -5412,7 +5412,7 @@ public sealed unsafe partial class DacDbiImpl : IDacDbiInterface
                 throw new ArgumentException("Output pointers must not be null.");
 
             ContinuationInfo info = _target.Contracts.Object.GetContinuationInfo(new TargetPointer(continuationAddress));
-            *pDiagnosticIP = info.DiagnosticIP.Value;
+            *pDiagnosticIP = (nuint)info.DiagnosticIP.Value;
             *pNextContinuation = info.Next.Value;
             *pState = info.State;
         }
@@ -5423,7 +5423,7 @@ public sealed unsafe partial class DacDbiImpl : IDacDbiInterface
 #if DEBUG
         if (_legacy is not null)
         {
-            ulong diagnosticIPLocal;
+            nuint diagnosticIPLocal;
             ulong nextLocal;
             uint stateLocal;
             int hrLocal = _legacy.ParseContinuation(continuationAddress, &diagnosticIPLocal, &nextLocal, &stateLocal);
